requirements.txt
pandas
numpy
torch
matplotlib
python-dotenv
requests
newsapi-python
textblob
ftfy
git+https://github.com/amazon-science/chronos-forecasting.git



